// FrontEnd/src/app/models/GraficoBVCTop10/GraficoBVCTop10.ts
export class GraficoBVCTop10 {
	CodMarca: number;
	DescMarca: string;
	CorSiteMarca: string;
	ShareDesejo: number;
	Efeitos: number;
	Equity: number;
	BaseShareDesejo: number;
	BaseEfeitos: number;
	BaseEquity: number;
	BaseMinimaShareDesejo: string;
	BaseMinimaEfeitos: string;
	BaseMinimaEquity: string;
}